requirements.txt
selenium
webdriver_manager
ruamel.yaml
